refactor(e2e): convert FeesTest to a single-node harness
Move e2e_fees/ into single-node/fees/ (joining the fee-asset price-oracle
test) and reconcile FeesTest onto SingleNodeTestContext. FeesTest now extends
the base and builds its prover-backed environment with its own setup() opts,
then calls hydrateFromContext to reuse the base rollup / epoch-cache /
chain-monitor / node-tracking / teardown machinery, keeping only the fee/gas
domain setup (FPC, fee juice, banana token) on the harness. The block-based
catchUpProvenChain is preserved verbatim. Adds the .parallel suffix to the
multi-it files and repoints the .test_patterns.yml fees entries.
