+++ .test_patterns.yml
@@ -108,7 +108,7 @@ tests:
     skip: true
     owners:
       - *alex
-  - regex: "src/e2e_fees/private_payments"
+  - regex: "src/single-node/fees/private_payments"
     owners:
       - *phil
   - regex: "src/e2e_token_contract/access_control"
@@ -364,7 +364,7 @@ tests:
     owners:
       - *lasse
 
-  - regex: "yarn-project/end-to-end/scripts/run_test.sh simple src/e2e_fees/gas_estimation.test.ts"
+  - regex: "yarn-project/end-to-end/scripts/run_test.sh simple src/single-node/fees/gas_estimation.parallel.test.ts"
     error_regex: "timeout: sending signal TERM to command 'bash'"
     owners:
       - *alex
